+++ Plugins/RHI/OpenGL/src/OpenGLShader.cpp
@@ -1,50 +1,138 @@
-#include "plugin_opengl/OpenGLShader.hpp"
+#include "opengl/OpenGLShader.hpp"
 #include "engine/core/Log.hpp"
 
-#include <glad/glad.h>
+#include <cstring>
+#include <filesystem>
+#include <fstream>
 #include <vector>
 
 namespace Leon {
 
+    static GLenum ShaderTypeFromString(const std::string& InType) {
+        if (InType == "vertex")
+            return GL_VERTEX_SHADER;
+        if (InType == "fragment" || InType == "pixel")
+            return GL_FRAGMENT_SHADER;
+
+        LE_CORE_ASSERT(false, "Unknown shader type!");
+        return 0;
+    }
+
+    FOpenGLShader::FOpenGLShader(const std::string& InFilePath) {
+        std::string source = ReadFile(InFilePath);
+        auto shaderSources = PreProcess(source);
+        Compile(shaderSources);
+
+        // Extract name from filepath (e.g. "Assets/Shaders/DirectionalLit.glsl" -> "DirectionalLit")
+        std::filesystem::path path = InFilePath;
+        m_Name = path.stem().string();
+    }
+
     FOpenGLShader::FOpenGLShader(const std::string& InName, const std::string& InVertexSrc,
                                  const std::string& InFragmentSrc)
         : m_Name(InName) {
-        GLuint vertexShader = CompileShader(GL_VERTEX_SHADER, InVertexSrc);
-        GLuint fragmentShader = CompileShader(GL_FRAGMENT_SHADER, InFragmentSrc);
+        std::unordered_map<GLenum, std::string> sources;
+        sources[GL_VERTEX_SHADER] = InVertexSrc;
+        sources[GL_FRAGMENT_SHADER] = InFragmentSrc;
+        Compile(sources);
+    }
 
-        m_RendererID = glCreateProgram();
-        glAttachShader(m_RendererID, vertexShader);
-        glAttachShader(m_RendererID, fragmentShader);
-        glLinkProgram(m_RendererID);
+    FOpenGLShader::~FOpenGLShader() {
+        if (m_RendererID) {
+            glDeleteProgram(m_RendererID);
+        }
+    }
+
+    std::string FOpenGLShader::ReadFile(const std::string& InFilePath) {
+        std::string result;
+        std::ifstream in(InFilePath, std::ios::in | std::ios::binary);
+        if (in) {
+            in.seekg(0, std::ios::end);
+            size_t size = in.tellg();
+            if (size != -1) {
+                result.resize(size);
+                in.seekg(0, std::ios::beg);
+                in.read(&result[0], size);
+            } else {
+                LE_CORE_ERROR("Could not read from file '{0}'", InFilePath);
+            }
+        } else {
+            LE_CORE_ERROR("Could not open file '{0}'", InFilePath);
+        }
+        return result;
+    }
+
+    std::unordered_map<GLenum, std::string> FOpenGLShader::PreProcess(const std::string& InSource) {
+        std::unordered_map<GLenum, std::string> shaderSources;
+
+        const char* typeToken = "#type";
+        size_t typeTokenLength = strlen(typeToken);
+        size_t pos = InSource.find(typeToken, 0);
+
+        while (pos != std::string::npos) {
+            size_t eol = InSource.find_first_of("\r\n", pos);
+            LE_CORE_ASSERT(eol != std::string::npos, "Syntax error in shader preprocessor line");
+            size_t begin = pos + typeTokenLength + 1;
+            std::string type = InSource.substr(begin, eol - begin);
+
+            // Trim any trailing whitespace
+            while (!type.empty() && (type.back() == ' ' || type.back() == '\t' || type.back() == '\r')) {
+                type.pop_back();
+            }
+
+            size_t nextLinePos = InSource.find_first_not_of("\r\n", eol);
+            pos = InSource.find(typeToken, nextLinePos);
+
+            shaderSources[ShaderTypeFromString(type)] = (pos == std::string::npos)
+                                                            ? InSource.substr(nextLinePos)
+                                                            : InSource.substr(nextLinePos, pos - nextLinePos);
+        }
+
+        return shaderSources;
+    }
+
+    void FOpenGLShader::Compile(const std::unordered_map<GLenum, std::string>& InShaderSources) {
+        GLuint program = glCreateProgram();
+        std::vector<GLenum> glShaderIDs;
+        glShaderIDs.reserve(InShaderSources.size());
+
+        for (auto& kv : InShaderSources) {
+            GLenum type = kv.first;
+            const std::string& source = kv.second;
+
+            GLuint shader = CompileShader(type, source);
+            if (shader) {
+                glAttachShader(program, shader);
+                glShaderIDs.push_back(shader);
+            }
+        }
+
+        glLinkProgram(program);
 
         GLint isLinked = 0;
-        glGetProgramiv(m_RendererID, GL_LINK_STATUS, &isLinked);
+        glGetProgramiv(program, GL_LINK_STATUS, &isLinked);
         if (isLinked == GL_FALSE) {
             GLint maxLength = 0;
-            glGetProgramiv(m_RendererID, GL_INFO_LOG_LENGTH, &maxLength);
+            glGetProgramiv(program, GL_INFO_LOG_LENGTH, &maxLength);
 
             std::vector<GLchar> infoLog(maxLength);
-            glGetProgramInfoLog(m_RendererID, maxLength, &maxLength, &infoLog[0]);
+            glGetProgramInfoLog(program, maxLength, &maxLength, &infoLog[0]);
 
-            glDeleteProgram(m_RendererID);
-            glDeleteShader(vertexShader);
-            glDeleteShader(fragmentShader);
+            glDeleteProgram(program);
+            for (auto id : glShaderIDs) {
+                glDeleteShader(id);
+            }
 
-            LE_CORE_ERROR("Shader link failure in \"{0}\": {1}", m_Name, infoLog.data());
+            LE_CORE_ERROR("Shader link failure in \"{0}\":\n{1}", m_Name, infoLog.data());
             return;
         }
 
-        glDetachShader(m_RendererID, vertexShader);
-        glDetachShader(m_RendererID, fragmentShader);
-        glDeleteShader(vertexShader);
-        glDeleteShader(fragmentShader);
-    }
+        for (auto id : glShaderIDs) {
+            glDetachShader(program, id);
+            glDeleteShader(id);
+        }
 
-    FOpenGLShader::FOpenGLShader(const std::string& InVertexSrc, const std::string& InFragmentSrc)
-        : FOpenGLShader("UnnamedShader", InVertexSrc, InFragmentSrc) {}
-
-    FOpenGLShader::~FOpenGLShader() {
-        glDeleteProgram(m_RendererID);
+        m_RendererID = program;
     }
 
     unsigned int FOpenGLShader::CompileShader(unsigned int InType, const std::string& InSource) {
@@ -64,8 +152,8 @@ namespace Leon {
 
             glDeleteShader(shader);
 
-            LE_CORE_ERROR("Shader compilation failure ({0}): {1}", (InType == GL_VERTEX_SHADER ? "VERTEX" : "FRAGMENT"),
-                          infoLog.data());
+            LE_CORE_ERROR("Shader compilation failure ({0}):\n{1}",
+                          (InType == GL_VERTEX_SHADER ? "VERTEX" : "FRAGMENT"), infoLog.data());
             return 0;
         }
 
@@ -86,7 +174,7 @@ namespace Leon {
 
         int location = glGetUniformLocation(m_RendererID, InName.c_str());
         if (location == -1)
-            LE_CORE_WARN("Uniform '{0}' not found!", InName);
+            LE_CORE_WARN("Uniform '{0}' not found in shader \"{1}\"!", InName, m_Name);
 
         m_UniformLocationCache[InName] = location;
         return location;